prototype_history/protoype11_fix/backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
sqlalchemy>=2.0.0
pydantic>=2.0.0
apscheduler>=3.10.0
python-dotenv>=1.0.0